Fix style Signed-off-by: Xin Wang <xin1.wang@intel.com>

+++ samples/cpp/benchmark_app/inputs_filling.cpp
@@ -254,7 +254,7 @@ ov::Tensor create_tensor_random(const benchmark_app::InputInfo& inputInfo,
     auto tensor = ov::Tensor(inputInfo.type, inputInfo.dataShape);
     auto data = tensor.data<T>();
 
-    std::mt19937 gen(std::chrono::system_clock::now().time_since_epoch().count());
+    std::mt19937 gen(static_cast<unsigned int>(std::chrono::system_clock::now().time_since_epoch().count()));
     uniformDistribution<T2> distribution(rand_min, rand_max);
     for (size_t i = 0; i < tensor_size; i++) {
         data[i] = static_cast<T>(distribution(gen));
